+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 
 - Add `onlyGenerateCoverageForSpecifiedTargets` parameter to `TestAction` https://github.com/tuist/XcodeProj/pull/473 by @kateinoigakukun
+- Added support for `PBXTargetDependency.product` https://github.com/tuist/XcodeProj/pull/481 by @yonaskolb
 
 ### Fixed